+++ Documentation/Changelog.md
@@ -10,12 +10,16 @@
     - `strict_tmp_tables=true`
     - `transaction_replay_timeout=30s`
 
-* MariaDBMonitor now requires MariaDB Server 10.4 or newer for failover/
+* MariaDB-Monitor now requires MariaDB Server 10.4 or newer for failover/
   switchover. Server 10.3 is end of life.
 
 * Server setting `private_address` added. Used for detecting and setting up
   replication. See
   [MariaDB Monitor documentation](Monitors/MariaDB-Monitor.md#private_address)
+  for more information.
+
+* MariaDB-Monitor can set the memory used by Mariabackup's prepare stage.
+  [MariaDB Monitor documentation](Monitors/MariaDB-Monitor.md#mariabackup_use_memory)
   for more information.
 
 * Pam authentication now always checks account status. Previously, this check